[fix] Resolve Size Definition Errors

+++ include/types.h
@@ -12,12 +12,12 @@
 #define TYPE_H
 
 typedef unsigned long long u64;
-typedef unsigned long      u32;
+typedef unsigned int       u32;
 typedef unsigned short     u16;
 typedef unsigned char      u8;
 
 typedef long long  i64;
-typedef long       i32;
+typedef int        i32;
 typedef short      i16;
 typedef char       i8;
